+++ app/controllers/admin/service_contacts_controller.rb
@@ -13,10 +13,13 @@ class Admin
     def update
       @contact = Contact.find(params[:id])
       @service = Service.find(params[:service_id])
+      location = @service.location
+
+      authorize location
 
       if @contact.update(contact_params)
         flash[:notice] = 'Contact was successfully updated.'
-        redirect_to [:admin, @service.location, @service, @contact]
+        redirect_to [:admin, location, @service, @contact]
       else
         render :edit
       end
@@ -33,9 +36,12 @@ class Admin
     def create
       @service = Service.find(params[:service_id])
       @contact = @service.contacts.new(contact_params)
+      location = @service.location
+
+      authorize location
 
       if @contact.save
-        redirect_to admin_location_service_path(@service.location, @service),
+        redirect_to admin_location_service_path(location, @service),
                     notice: "Contact '#{@contact.name}' was successfully created."
       else
         render :new
@@ -44,8 +50,13 @@ class Admin
 
     def destroy
       contact = Contact.find(params[:id])
+      service = contact.service
+      location = service.location
+
+      authorize location
+
       contact.destroy
-      redirect_to admin_location_service_path(contact.service.location, contact.service),
+      redirect_to admin_location_service_path(location, service),
                   notice: "Contact '#{contact.name}' was successfully deleted."
     end
